+++ src/skipper/program.hpp
@@ -90,15 +90,15 @@ namespace skipper
         /// @copydoc add_command
         /// @param validate Predicate that validates the input for the command
         /// @param convert function that converts the input string
-        template<typename Type, typename Predicate>
+        template<typename Type, typename Predicate = any<Type>>
         void add_command(const std::string key,
                          std::string description,
                          const std::function<void(const Type)>& call,
-                         const Predicate& validate,
+                         const Predicate& validate = any<Type>(),
                          std::function<const Type(std::istream&)> convert =
                             default_convert<Type>)
         {
-            assert(!m_commands.count(key));
+            assert(!m_commands.count(key) && "Duplicated command");
 
             auto set_command = [=]()
             {