+++ NEWS.rst
@@ -6,6 +6,11 @@ every change, see the Git log.
 
 Latest
 ------
+* Major: Changed constructor program(...) -> program()
+* Major: Changed run(bool print_help) -> run()
+* Major: Added depedendcy on kw
+* Minor: Added optional arguments in, out, set_print_help_on_run,
+  ready_indicator
 * Minor: Validator defaults to Any input, if no validator is provided.
 
 1.2.0