+++ NEWS.rst
@@ -10,6 +10,7 @@ Latest
 * Minor: Validator defaults to Any input, if no validator is provided.
 * Minor: Added set_print_help(bool) to turn on/off help print on run.
 * Minor: Added set_ready_indicator to provide custom ready indicator.
+* Minor: Allow the exit key to be set via set_exit_key.
 
 1.2.0
 -----